requirements.txt
click==8.1.7
ipykernel==6.25.2
ipython==8.12.3
isort==5.12.0
matplotlib==3.7.3
numpy==1.24.3
pandas==2.0.3
Pillow==10.0.1
PyYAML==6.0.1
scikit-learn==1.3.1
tensorflow==2.13.0
tensorflow-addons==0.21.0
tensorflow-io==0.34.0
petname==2.2